Update onoff capability listener to parse value
The onoff capability listener now takes the expected onoff value, as
given by Homey, into account of whether to turn devices on or off.
Previously the `onCapabilityOnoff` function switches the device state,
which in turn results in devices turning on when askes to turn off
twice.

Now the expected value is taken into account whether to turn the devices
on or off.

Fix #73