Fix executing Toggle mode if state is already correct
For devices that are turned on/off via Toggle it is important that this
is only executed if device has to change state.

Currently if device is off and it is requested to turn off, toggle mode
is still executed. This could result in device actually being turned on.

Fix is to disable executing Toggle mode if the desired state is equal to
the current state of the device.